Automate API documentation generation and deployment
This commit adds two GitHub Actions workflows: docs and docs-versioned.

Workflow "docs" generates API documentation on each push to main and
deploy it to GitHub Pages.  The repository needs to have GitHub Pages
enabled and GitHub Actions setup as its input.

Workflow "docs-versioned" also generates the documentation, but on
creation of tags of v[0-9]+.[0-9]+.[0-9]+ format, and then copies
the result to portals-project/portals-api repository in a directory
named the same way as tag.